Lamp shade based on Ikea Sekond
Lampshade based on Ikea Sekond lamp profile.
Was made for the Prusa comp "Let there be light" but due to sickness I didn't upload it on time.
Same print 8 times to make the full lampshade.
Made in Fusion 360 and rendered in it. I choose ASA , because light sources usually produce a certain level of heat. ASA or PC Blend may be a better choice in this case. Also I strongly recommend a led light bulb!